fix: subtitle sync — native track + media-time timestamps
- Replace HLS.js subtitle loading with native <track> element pointing to subs.vtt
- Refresh track src every 5s to pick up new chunks
- Use chunk_start_time (media-time) instead of shifted_start (wall-clock) for VTT timestamps
- Remove DISCONTINUITY tags from subtitle playlist
- Remove pipeline_delay compensation from subtitle timestamps (was causing 10s delay)
- Remove HLS.js subtitle event handlers
- SubtitleGenerator: media_seq=0 always, absolute VTT timestamps

+++ modules/outputs/hls_output.py
@@ -379,7 +379,7 @@ class HLSOutput(OutputSink):
 
         if "nvenc" in encoder:
             fps = self._encoder_config.video_fps or 25
-            gop_frames = int(round(fps * self._segment_duration))
+            gop_frames = round(fps * self._segment_duration)
             cmd.extend(["-preset", preset, "-g", str(gop_frames), "-keyint_min", str(gop_frames), "-no-scenecut", "1"])
         elif encoder == "libx264":
             cmd.extend(["-preset", preset])
